Support multiple private zones w/ same zone name
Multiple private zones with the same zone name (i.e. for testing, staging,
production) are exported to the same file when using the `--split`
option.

By adding the `vpc_id` to the `route_file_basename`, multiple private
zones with the same zone name can be correctly exported to separate
files when using the `--split` option.