dsl-tester: A(Alias) fake diff fixed
Thanks to the non-standard Alias record type that has been introduced by
AWS Route 53 one could get fake diffs while running `--test`.
This had no effect on the actual `--apply` command since nothing would
change but could mislead the user that the record is going to be updated.

The root cause is the lack of the actual TTL field on the Alias records so
when the `dsl-tester` tried to check if the actual and expected values
match it showed that they are different.

Since the A (ipv4) record (returned by the query) has TTL and the Alias has not.

Solution:

I've introduced a `check_ttl` variable that sets if the TTL values
should be checked or not. The code simply disables ttl check if it's an
Alias record.

Signed-off-by: Gergő Nagy <contact@gergonagy.com>